requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.16.0
python-dotenv>=1.0.0
